requirements.txt
streamlit
torch
torchvision
pandas
numpy
Pillow
